Probe for new noload RTT as a function of current limit
Right now, the noload RTT is computed when it reaches zero. What this
means is that (using default settings, with a windowed limiter), if I
have a concurrency limit of 10 then I'll reset my noload RTT between
every 1500 and 3000 requests.

This is fine, but deals poorly with a limit that dramatically reduces.
For example, if I have a concurrency limit of 100, but then system load
goes up and my permits dramatically decrease to 10, I will still have to
wait for between 15k and 30k requests in order to reset any noload based
issues.

This PR changes it to be based on the current rate limit. So if for
example I compute that I need to wait 17k requests before I reset my
noload, then if my concurrency drops down by a factor of 10, it'll
become 1.7k requests.